Extracted we_wind.h from edit.h
