+++ WeSyntax.h
@@ -1,0 +1,6 @@
+#ifndef WESYNTAX_H
+#define WESYNTAX_H
+
+#include "we_wind.h"
+
+#endif
