+++ we_wind.h
@@ -36,4 +36,39 @@
 /*	is the coordinate for the lower right corner*/
 /******************************************************************************/
 
+
+/*   we_wind.c   */
+int e_error(char *text, int sw, FARBE *f);
+int e_message(int sw, char *str, FENSTER *f);
+void e_firstl(FENSTER *f, int sw);
+int e_pr_filetype(FENSTER *f);
+PIC *e_open_view(int xa, int ya, int xe, int ye, int col, int sw);
+int e_close_view(PIC *pic, int sw);
+void e_pr_line(int y, FENSTER *f);
+void e_std_rahmen(int xa, int ya, int xe, int ye, char *name, int sw,
+  int frb, int fes);
+void e_ed_rahmen(FENSTER *f, int sw);
+int e_schirm(FENSTER *f, int sw);
+int e_size_move(FENSTER *f);
+PIC *e_std_kst(int xa, int ya, int xe, int ye, char *name, int sw, int fr,
+  int ft, int fes);
+PIC *e_ed_kst(FENSTER *f, PIC *pic, int sw);
+int e_close_window(FENSTER *f);
+void e_switch_window(int num, FENSTER *f);
+int e_ed_zoom(FENSTER *f);
+int e_ed_cascade(FENSTER *f);
+int e_ed_tile(FENSTER *f);
+int e_ed_next(FENSTER *f);
+int e_mess_win(char *header, char *str, PIC **pic, FENSTER *f);
+PIC *e_change_pic(int xa, int ya, int xe, int ye, PIC *pic, int sw, int frb);
+struct dirfile *e_add_df(char *str, struct dirfile *df);
+int e_schr_nchar_wsv(char *str, int x, int y, int n, int max, int col,
+  int csw);
+int e_schr_lst_wsv(char *str, int xa, int ya, int n, int strlen, int ft,
+  int fz, struct dirfile **df, FENSTER *f);
+int e_rep_win_tree(ECNT *cn);
+int e_opt_sec_box(int xa, int ya, int num, OPTK *opt, FENSTER *f, int sw);
+int e_close_buffer(BUFFER *b);
+int e_list_all_win(FENSTER *f);
+
 #endif
